ImageMetadataKeys, ImageTypes: move to model namespace
