+++ src/Snowflake.Framework.Primitives/Model/Records/File/ImageMetadataKeys.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Snowflake.Caching
+namespace Snowflake.Model.Records.File
 {
     /// <summary>
     /// The image metadata key
